+++ Main/GameEvents/InputGameEvent/InputSystemGameEvent/InputSystemGameEventConfig.cs
@@ -1,11 +1,13 @@
 using System;
 using Sirenix.OdinInspector;
+using UnityEngine;
 using UnityEngine.InputSystem;
 using VMFramework.Core;
 using VMFramework.OdinExtensions;
 
 namespace VMFramework.GameEvents
 {
+    [Serializable]
     public class InputSystemGameEventConfig : GameEventConfig
     {
         public override Type GameItemType => typeof(InputSystemGameEvent);
@@ -13,6 +15,10 @@ namespace VMFramework.GameEvents
         [TabGroup(TAB_GROUP_NAME, BASIC_CATEGORY)]
         [InputActionID]
         public Guid inputActionID;
+
+        [SerializeField]
+        [HideInInspector]
+        private string nativeInputActionID;
         
         public InputAction InputAction { get; private set; }
 
@@ -26,6 +32,12 @@ namespace VMFramework.GameEvents
             {
                 UnityEngine.Debug.LogError($"Input action with ID {inputActionID} not found.");
             }
+        }
+
+        protected override void OnPrepareNativeSerializationMigration()
+        {
+            base.OnPrepareNativeSerializationMigration();
+            nativeInputActionID = inputActionID.ToString("D");
         }
     }
 }