+++ Main/UI/Panel/Popups/PopupText/UGUIPopupTextConfig.cs
@@ -2,6 +2,7 @@
 
 namespace VMFramework.UI
 {
+    [Serializable]
     public class UGUIPopupTextConfig : UGUIPanelConfig, IPopupTextConfig
     {
         public override Type GameItemType => typeof(UGUIPopupText);