+++ Main/UI/Panel/UIPanels/UGUIPanel/UGUIPanelConfig.cs
@@ -2,6 +2,7 @@
 
 namespace VMFramework.UI
 {
+    [Serializable]
     public class UGUIPanelConfig : UIPanelConfig, IUGUIPanelConfig
     {
         protected const string UGUI_PANEL_CATEGORY = "UGUI";