+++ MapExtension/TileBaseConfig/DefaultTileBaseConfig/DefaultTileBaseConfig.cs
@@ -1,8 +1,10 @@
 ﻿using Sirenix.OdinInspector;
+using System;
 using UnityEngine.Tilemaps;
 
 namespace VMFramework.Maps
 {
+    [Serializable]
     public sealed partial class DefaultTileBaseConfig : TileBaseConfig
     {
         [TabGroup(TAB_GROUP_NAME, BASIC_CATEGORY)]