+++ MapExtension/TileBaseConfig/TileBaseConfig/TileBaseConfig.cs
@@ -1,8 +1,10 @@
+using System;
 using UnityEngine.Tilemaps;
 using VMFramework.GameLogicArchitecture;
 
 namespace VMFramework.Maps 
 {
+    [Serializable]
     public abstract partial class TileBaseConfig : GamePrefab, ITileBaseConfig
     {
         public override string IDSuffix => "tile_base";